- size_t related bug fixes, addressing to "unsigned vs. signed types" issues   (Ralf)

+++ apt-pkg/contrib/strutl.cc
@@ -968,7 +968,7 @@ void ioprintf(ostream &out,const char *format,...)
 char *safe_snprintf(char *Buffer,char *End,const char *Format,...)
 {
    va_list args;
-   unsigned long Did;
+   int Did;
 
    va_start(args,Format);
